[ refactor ] filter tag type convert 관련 로직 분리
