+++ src/core/atom/slider.ts
@@ -1,6 +1,6 @@
-import { atom } from "recoil";
+import { atom, DefaultValue } from "recoil";
 
-import { persistedFilterTagsStateEffect } from "./filterTagsStateHandlers";
+import { convertArrToSet, convertSetToArr } from "../../util/cardCollection/filterTypeConverter";
 
 enum StateType {
   SLIDER_IDX = "slideridx",
@@ -21,5 +21,21 @@ export const isSliderDownState = atom({
 export const filterTagsState = atom<{ tags: Set<string>; intimacy: number[]; isActive: boolean }>({
   key: StateType.FILTER_TAGS,
   default: { tags: new Set(), intimacy: [0], isActive: false },
-  effects: [persistedFilterTagsStateEffect],
+  effects: [
+    ({ setSelf, onSet }) => {
+      const savedValue = sessionStorage.getItem(StateType.FILTER_TAGS);
+
+      if (savedValue !== null) {
+        setSelf(convertArrToSet(JSON.parse(savedValue)));
+      }
+
+      onSet((newValue) => {
+        if (newValue instanceof DefaultValue) {
+          sessionStorage.removeItem(StateType.FILTER_TAGS);
+        } else {
+          sessionStorage.setItem(StateType.FILTER_TAGS, JSON.stringify(convertSetToArr(newValue)));
+        }
+      });
+    },
+  ],
 });
